orchestrate-server/requirements.txt
asyncio
boto3
Flask==2.0.2
flask-cors==3.0.10
# grouphug
huggingface_hub
joblib
matplotlib
miditoolkit==0.1.16
mido==1.3.0
midi2audio
music21==8.3.0
numpy==1.20.3
openai==0.28
pandas
parmap
pretty-midi
pydantic
python-dotenv==0.19.2
PyYAML==6.0
scipy
split-folders
torch
transformers
Werkzeug==2.2.2
waitress # for production
yacs
